5-manage_your_vnf_with_an_helm_chart_based_execution_environment/grpc/osm_ee/requirements.txt
grpcio-tools
grpclib
PyYAML
asyncssh
ansible
python-dotenv
cryptography
